+++ lib/layer.dart
@@ -1,0 +1,2 @@
+library web.layer;
+export 'src/layer.dart';