+++ lib/src/cmd.dart
@@ -1,3 +1,6 @@
+import 'dart:async' show Future;
+
+typedef Future CommandHandler(Command command);
 
 
 class Command{
@@ -7,4 +10,6 @@ class Command{
   Command(this.command,{this.args}){
     if (args == null) args = const {};
   }
+
+  String toString() => "Command $command -> $args";
 }