+++ lib/web.dart
@@ -1,6 +1,7 @@
+library web;
 export "src/web.dart";
 export "src/request.dart";
 export "src/config.dart";
 export "src/context.dart";
-export "src/layer.dart";
 export "src/plugin.dart";
+export 'src/logging.dart' show getLogger;