+++ lib/webart.dart
@@ -1,1 +1,2 @@
+library web;
 export './web.dart';