+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # CHANGELOG
 
-## 0.2.0-alpha.7
-- Will be done later
+## 0.2.0-beta.0
+- New `MessageChannel`, `ChannelSession`, `ChannelSeesionMessage` and `Command`
+- Return HTTP error 404 when handler not found
+- Router will be fast if uri path is static (for example '/about')
+- Builtin Router now work as a Plugin by `RoutingPlugin`, it also use third party router if given
+- A dirty fix for uri template "" match all uri
+- Package structure changed, all of Layer moved to single 'package:webart/layer.dart'
+- Small change for example.
+- Code opz.,remove unused imports and functions
+- Mark EvevntBus as deprecated, it will be removed when 0.2.0 released.
 
 ## 0.2.0-alpha.6
 - Remove `LoggingLayer`